test(examples): convert unordered-collections example to YAML
Replace the TypeScript/Node.js unordered-collections example program with a
runtime: yaml program so the integration test no longer requires a JS/TS
package or yarn linking. The config-driven ordering flip is expressed with an
integer config selecting between two authored orderings via fn::select.

Move TestUnorderedCollectionsRefresh (and its helpers) out of the
nodejs-tagged test file into examples_test.go, which has no build tag, and
drop the YarnLink option since YAML needs no per-language toolchain.
