+++ examples/examples_nodejs_test.go
@@ -97,87 +97,6 @@ func TestRefreshChanges(t *testing.T) {
 	assert.Equal(t, 1, len(*updateSummary))
 }
 
-func TestUnorderedCollectionsRefresh(t *testing.T) {
-	_ = getEnvRegion(t)
-	cwd := getCwd(t)
-	test := pulumitest.NewPulumiTest(t, filepath.Join(cwd, "unordered-collections"),
-		opttest.LocalProviderPath("aws-native", filepath.Join(cwd, "..", "bin")),
-		opttest.YarnLink("@pulumi/aws-native"),
-	)
-	defer test.Destroy(t)
-
-	test.Up(t)
-	test.Refresh(t)
-
-	deploymentJSON, err := test.ExportStack(t).Deployment.MarshalJSON()
-	require.NoError(t, err)
-	var deployment apitype.DeploymentV3
-	require.NoError(t, json.Unmarshal(deploymentJSON, &deployment))
-
-	var role, taskDefinition *apitype.ResourceV3
-	for i := range deployment.Resources {
-		switch deployment.Resources[i].Type {
-		case "aws-native:iam:Role":
-			role = &deployment.Resources[i]
-		case "aws-native:ecs:TaskDefinition":
-			taskDefinition = &deployment.Resources[i]
-		}
-	}
-	require.NotNil(t, role, "expected IAM Role in exported deployment")
-	require.NotNil(t, taskDefinition, "expected ECS TaskDefinition in exported deployment")
-
-	// IAM returns inline policy names alphabetically, proving that refresh saw
-	// AWS order while the program and checkpoint retain the authored order.
-	assert.Equal(t, []string{"z-policy", "a-policy"}, rolePolicyNames(t, role.Inputs["policies"]))
-	assert.Equal(t, []string{"a-policy", "z-policy"}, rolePolicyNames(t, role.Outputs["policies"]))
-
-	// ECS preserves the authored environment order. Reverse only the program
-	// order after refresh to exercise nested unordered Diff against live state.
-	assert.Equal(t, []string{"Z_VAR", "A_VAR"}, taskEnvironmentNames(t, taskDefinition.Inputs))
-	assert.Equal(t, []string{"Z_VAR", "A_VAR"}, taskEnvironmentNames(t, taskDefinition.Outputs))
-	test.SetConfig(t, "reverseEnvironment", "true")
-
-	previewResult := test.Preview(t)
-	assertpreview.HasNoChanges(t, previewResult)
-}
-
-func rolePolicyNames(t *testing.T, value any) []string {
-	t.Helper()
-	policies, ok := value.([]interface{})
-	require.True(t, ok, "expected policies array, got %T", value)
-
-	names := make([]string, len(policies))
-	for i, rawPolicy := range policies {
-		policy, ok := rawPolicy.(map[string]interface{})
-		require.True(t, ok, "expected policy object, got %T", rawPolicy)
-		name, ok := policy["policyName"].(string)
-		require.True(t, ok, "expected policyName string, got %T", policy["policyName"])
-		names[i] = name
-	}
-	return names
-}
-
-func taskEnvironmentNames(t *testing.T, properties map[string]interface{}) []string {
-	t.Helper()
-	containers, ok := properties["containerDefinitions"].([]interface{})
-	require.True(t, ok, "expected containerDefinitions array, got %T", properties["containerDefinitions"])
-	require.Len(t, containers, 1)
-	container, ok := containers[0].(map[string]interface{})
-	require.True(t, ok, "expected container definition object, got %T", containers[0])
-	environment, ok := container["environment"].([]interface{})
-	require.True(t, ok, "expected environment array, got %T", container["environment"])
-
-	names := make([]string, len(environment))
-	for i, rawVariable := range environment {
-		variable, ok := rawVariable.(map[string]interface{})
-		require.True(t, ok, "expected environment object, got %T", rawVariable)
-		name, ok := variable["name"].(string)
-		require.True(t, ok, "expected environment name string, got %T", variable["name"])
-		names[i] = name
-	}
-	return names
-}
-
 func TestPreviewStableProperties(t *testing.T) {
 	cwd := getCwd(t)
 	options := []opttest.Option{
