--- examples/unordered-collections/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "outDir": "bin",
-    "target": "es2016",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "pretty": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitReturns": true,
-    "forceConsistentCasingInFileNames": true
-  },
-  "files": [
-    "index.ts"
-  ]
-}
